backend/requirements.txt
flask
flask_cors
pymongo
datetime
python-dotenv
